+++ R/00-bregr-package.R
@@ -4,11 +4,12 @@
 ## usethis namespace: start
 #' @import vctrs
 #' @import cli
-#' @importFrom survival coxph
+#' @importFrom mirai .progress
+#' @importFrom survival Surv survdiff survfit coxph
 #' @importFrom lifecycle deprecated
 #' @importFrom rlang .data .env
 #' @importFrom utils combn packageDescription packageVersion
-#' @importFrom stats cor
+#' @importFrom stats cor pchisq quantile predict sd
 #' @rawNamespace if (getRversion() < "4.3.0") importFrom("S7", "@")
 ## usethis namespace: end
 NULL